+++ docs/alphadia/outputtransform.rst
@@ -1,7 +1,7 @@
 outputtransform
 ===============
 
-The outputtransform module contains functionality to perform the consensus workflow.
+The outputtransform package contains functionality to perform the consensus workflow.
 Individual search workflows, part of a joined search plan, are combined.
 
 This includes:
